Update ios.yml
Refined the workflow YAML for the iOS project, addressing issues in variable assignments and project/workspace file determination. Fixed the variable assignment for filetype_parameter in the "Build" step and corrected the determination of the project/workspace file based on the existence of the .xcworkspace file. Also replaced backticks with $(...) syntax for command substitution, enhancing consistency and readability throughout the workflow YAML.